# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WanderLens Travel Memory App API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Photo {
  id: number;
  filename: string;
  originalName: string;
  mimeType: string;
  /** @nullable */
  fileHash?: string | null;
  cloudinaryPublicId: string;
  cloudinaryUrl: string;
  /** @nullable */
  lat?: number | null;
  /** @nullable */
  lng?: number | null;
  /** @nullable */
  takenAt?: string | null;
  /** @nullable */
  tripId?: number | null;
  createdAt: string;
}

export interface PhotoUpload {
  file: Blob;
}

export interface Trip {
  id: number;
  name: string;
  startDate: string;
  endDate: string;
  /** @nullable */
  coverPhotoId?: number | null;
  /** @nullable */
  coverPhotoPath?: string | null;
  /** @nullable */
  coverCloudinaryUrl?: string | null;
  /** @nullable */
  locationName?: string | null;
  /** @nullable */
  centerLat?: number | null;
  /** @nullable */
  centerLng?: number | null;
  photoCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface TripPatch {
  name?: string;
  /** @nullable */
  coverPhotoId?: number | null;
  /** @nullable */
  locationName?: string | null;
}

export interface TripMapPin {
  id: number;
  name: string;
  centerLat: number;
  centerLng: number;
  photoCount: number;
  startDate: string;
  endDate: string;
  /** @nullable */
  coverPhotoPath?: string | null;
  /** @nullable */
  coverCloudinaryUrl?: string | null;
  /** @nullable */
  locationName?: string | null;
}

export interface Stats {
  totalPhotos: number;
  totalTrips: number;
  photosWithLocation: number;
  /** @nullable */
  earliestDate: string | null;
  /** @nullable */
  latestDate: string | null;
  totalCountries?: number;
}

export interface MergeTripsBody {
  sourceId: number;
}

export interface RegroupResult {
  tripsCreated: number;
  photosGrouped: number;
}

export type ListPhotosParams = {
tripId?: number;
};