# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: WanderLens Travel Memory App API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: photos
    description: Photo upload and management
  - name: trips
    description: Trip management
  - name: stats
    description: Dashboard statistics

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /photos:
    get:
      operationId: listPhotos
      tags: [photos]
      summary: List all photos
      parameters:
        - in: query
          name: tripId
          schema:
            type: integer
      responses:
        "200":
          description: List of photos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Photo"

  /photos/upload:
    post:
      operationId: uploadPhoto
      tags: [photos]
      summary: Upload a photo (JPEG, PNG, HEIC)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/PhotoUpload"
      responses:
        "201":
          description: Uploaded photo with EXIF data and trip assignment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Photo"
        "400":
          description: Invalid file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /photos/{id}:
    get:
      operationId: getPhoto
      tags: [photos]
      summary: Get a photo by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Photo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Photo"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deletePhoto
      tags: [photos]
      summary: Delete a photo
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /trips:
    get:
      operationId: listTrips
      tags: [trips]
      summary: List all trips ordered by date descending
      responses:
        "200":
          description: List of trips
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trip"

  /trips/map:
    get:
      operationId: getTripsMap
      tags: [trips]
      summary: All trips with coordinates and photo count for world map
      responses:
        "200":
          description: Trip map pins
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TripMapPin"

  /trips/{id}:
    get:
      operationId: getTrip
      tags: [trips]
      summary: Get a trip by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateTrip
      tags: [trips]
      summary: Update a trip (rename, change cover)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripPatch"
      responses:
        "200":
          description: Updated trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteTrip
      tags: [trips]
      summary: Delete a trip and its photos
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /trips/{id}/merge:
    post:
      operationId: mergeTrips
      tags: [trips]
      summary: Merge source trip into this trip (source is deleted)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MergeTripsBody"
      responses:
        "200":
          description: Updated target trip after merge
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
        "404":
          description: Trip not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /trips/{id}/photos:
    get:
      operationId: getTripPhotos
      tags: [trips]
      summary: Get all photos for a trip
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Photos in the trip
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Photo"
        "404":
          description: Trip not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /stats:
    get:
      operationId: getStats
      tags: [stats]
      summary: Dashboard summary stats
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stats"

  /photos/regroup:
    post:
      operationId: regroupPhotos
      tags: [photos]
      summary: Re-run trip grouping logic on all ungrouped photos
      responses:
        "200":
          description: Grouping result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegroupResult"

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    Photo:
      type: object
      required: [id, filename, originalName, mimeType, cloudinaryPublicId, cloudinaryUrl, createdAt]
      properties:
        id:
          type: integer
        filename:
          type: string
        originalName:
          type: string
        mimeType:
          type: string
        fileHash:
          type: ["string", "null"]
        cloudinaryPublicId:
          type: string
        cloudinaryUrl:
          type: string
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
        takenAt:
          type: ["string", "null"]
        tripId:
          type: ["integer", "null"]
        createdAt:
          type: string

    PhotoUpload:
      type: object
      required: [file]
      properties:
        file:
          type: string
          format: binary

    Trip:
      type: object
      required: [id, name, startDate, endDate, photoCount, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        coverPhotoId:
          type: ["integer", "null"]
        coverPhotoPath:
          type: ["string", "null"]
        coverCloudinaryUrl:
          type: ["string", "null"]
        locationName:
          type: ["string", "null"]
        centerLat:
          type: ["number", "null"]
        centerLng:
          type: ["number", "null"]
        photoCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string

    TripPatch:
      type: object
      properties:
        name:
          type: string
        coverPhotoId:
          type: ["integer", "null"]
        locationName:
          type: ["string", "null"]

    TripMapPin:
      type: object
      required: [id, name, centerLat, centerLng, photoCount, startDate, endDate]
      properties:
        id:
          type: integer
        name:
          type: string
        centerLat:
          type: number
        centerLng:
          type: number
        photoCount:
          type: integer
        startDate:
          type: string
        endDate:
          type: string
        coverPhotoPath:
          type: ["string", "null"]
        coverCloudinaryUrl:
          type: ["string", "null"]
        locationName:
          type: ["string", "null"]

    Stats:
      type: object
      required: [totalPhotos, totalTrips, photosWithLocation, earliestDate, latestDate]
      properties:
        totalPhotos:
          type: integer
        totalTrips:
          type: integer
        photosWithLocation:
          type: integer
        earliestDate:
          type: ["string", "null"]
        latestDate:
          type: ["string", "null"]
        totalCountries:
          type: integer

    MergeTripsBody:
      type: object
      required: [sourceId]
      properties:
        sourceId:
          type: integer

    RegroupResult:
      type: object
      required: [tripsCreated, photosGrouped]
      properties:
        tripsCreated:
          type: integer
        photosGrouped:
          type: integer